fix: remove setuptools-scm (#197)
## Reason for Change

- setuptools-scm and release-please don't play nice together. The [git
command
](https://setuptools-scm.readthedocs.io/en/latest/config/#setuptools_scm.git.DEFAULT_DESCRIBE)
command used by setuptools-scm is unable to parse the python-api version
if multiple tags are on the same commit. This causes pypi releases to
fail if they are released the same time as ontology-assets.

## Changes

- remove setuptool-scm from the python-api build chain. 
- use release-please to update the python version values in the API

## Testing steps
- Ran release-please. This is what a release would look like
https://github.com/chanzuckerberg/cellxgene-ontology-guide/pull/198
- verified unit tests pass

## Notes for Reviewer

### Pros
- reduces number of dependencies
- automated pypi releases will work reliably

### Cons
- When build locally on a dirty branch, the version of COG will still be
whatever the last release was. If a users does try to push this version
to pypi it will fail because that version should already be release.