feat: put out alpha release of 7.0.0 (#297)
## Reason for Change

when trying to install from git in single-cell-curation:
https://github.com/chanzuckerberg/single-cell-curation/pull/1443 it
looks like tests are failing
https://github.com/chanzuckerberg/single-cell-curation/actions/runs/17559487383/job/49871988778?pr=1443
because we can't find stuff that's in `ontology-assets`. since the git
install can only install from `cellxgene-ontology-guide/api/python`, it
doesn't include anything that's in `ontology-assets`

as far as i can tell i don't think we can pip install COG from git such
that we include `ontology-assets` since the pyproject.toml is all
located in `api/python`. it may be possible if we change the directory
structure of things but i think the simplest way to resolve this is just
to abandon the git install and put out an alpha release of COG and use
that in cellxgene-schema

## Changes

so this PR updates the 7.0.0 schema to 7.0.0-alpha, and i'll update the
autorelease PR that's generated to put forth an alpha release

## Testing steps

n/a

## Notes for Reviewer

---------

Co-authored-by: github-actions <github-actions@github.com>