feat: remove + character from regex [VC-4536]  (#313)
Small change to Regex:

the `+` character allows for one to many of the immediately previous
capture group (in this case, since no capture group is explicitly
stated, it just defaults to the previous character). This was resulting
in the regex expression yielding valid for things that our logic
dictates to be invalid. For example:

The regex `"^ZFA+:[0-9]+$"` would be valid for the following:
- `ZFA:1234`
- `ZFAAA:1234`
- `ZFAAAAAAAAAAAAAAAAAAAA:1234` 

Where as we should only expect it to be valid for the first one.
The change in this PR simply removes the `+` symbol where relevant

Co-authored-by: Kevin Liu <kliu@mac.lan>